completed footer, added responsiveness
